requirements.txt
pygame-ce>=2.5.5
torch>=2.1.0
numpy>=1.24.3
matplotlib>=3.7.2 